+++ code/midtown/memory/stack.h
@@ -44,10 +44,10 @@ ARTS_IMPORT void DebugLogInit(i32 arg1);
 ARTS_IMPORT void DebugLogShutdown();
 
 // 0x5204C0 | ?DoStackTraceback@@YAXHPAH@Z
-ARTS_EXPORT void DoStackTraceback(i32 depth, i32* frame);
+ARTS_EXPORT void DoStackTraceback(i32 depth, isize* frame);
 
 // 0x520670 | ?DumpStackTraceback@@YAXPAHH@Z
-ARTS_EXPORT void DumpStackTraceback(i32* frames, i32 count);
+ARTS_EXPORT void DumpStackTraceback(isize* frames, i32 count);
 
 // 0x5205B0 | ?LogStackTraceback@@YAHPAHH@Z
 ARTS_IMPORT i32 LogStackTraceback(i32* arg1, i32 arg2);
@@ -57,7 +57,7 @@ void LookupAddress(char* buffer, usize buflen, usize address);
 // 0x520130 | ?LookupAddress@@YAXPADH@Z
 // void LookupAddress(char* buffer, i32 address);
 
-i32 StackTraceback(i32 depth, i32* frames, i32 skipped);
+i32 StackTraceback(i32 depth, isize* frames, i32 skipped);
 
 // 0x520590 | ?StackTraceback@@YAXH@Z
 ARTS_EXPORT void StackTraceback(i32 depth);
