+++ code/midtown/stream/filestream.h
@@ -54,7 +54,7 @@ public:
     ARTS_EXPORT FileStream(i32 handle);
 
     // 0x5617F0 | ??0FileStream@@QAE@PAXHPAVFileSystem@@@Z
-    ARTS_EXPORT FileStream(void* buffer, i32 buffer_size, class FileSystem* file_system);
+    ARTS_EXPORT FileStream(void* buffer, isize buffer_size, class FileSystem* file_system);
 
     // 0x561B10 | ??_EFileStream@@UAEPAXI@Z
     // 0x561A60 | ??1FileStream@@UAE@XZ
@@ -73,7 +73,7 @@ public:
     ARTS_EXPORT i32 Open(const char* path, b32 read_only);
 
     // 0x561970 | ?RawRead@FileStream@@UAEHPAXH@Z
-    ARTS_EXPORT i32 RawRead(void* ptr, i32 size) override;
+    ARTS_EXPORT isize RawRead(void* ptr, isize size) override;
 
     // 0x5619B0 | ?RawSeek@FileStream@@UAEHH@Z
     ARTS_EXPORT i32 RawSeek(i32 pos) override;
@@ -85,7 +85,7 @@ public:
     ARTS_EXPORT i32 RawTell() override;
 
     // 0x561990 | ?RawWrite@FileStream@@UAEHPAXH@Z
-    ARTS_EXPORT i32 RawWrite(const void* ptr, i32 size) override;
+    ARTS_EXPORT isize RawWrite(const void* ptr, isize size) override;
 
     // 0x561950 | ?Stderr@FileStream@@QAEHXZ | unused
     ARTS_EXPORT i32 Stderr();
